Update to resolve issue with edit function when data source is being updated by an Observable
New settings 'confirm.resolve.skipEdit'=true will not perform any change to the data source after an edit has been performed. This change was done to resolve an issue when the data source is being updated by an Observable. This will allow for a better integration of the smart table to rxjs and ngrx.