Only update filter onChanged() if setFilter was the action
This fix is to accomodate slow servers, where a user could be mid-typing, and if the server then responds, their input gets cleared and reset to the filter the server last saw. This is partially due to the input debounce time. Since the app doesn't know the user wants to update the filter until they are done typing.

Fixes #681